Simplify sources of user metadata
Gafaelfawr supported a staggering variety of combinations of options
for where to get user metadata (name, email, UID, primary GID, and
group membership), including several varieties of partial LDAP
support. Simplify this to only three supported cases:

* GitHub (unchanged from before)
* OIDC with UIDs and GIDs from LDAP
* OIDC with LDAP but UIDs and GIDs from Firestore

(CILogon is just an OIDC implementation from Gafaelfawr's perspective.)
Drop all the code to get user metadata out of the OIDC token, except
for the username, and streamline the userinfo service.

Simplify the test configuration, collapse the LDAP tests into the
regular OIDC tests, and separate out the error tests.
