+++ changelog.d/20240502_133255_rra_DM_44136.md
@@ -1,0 +1,5 @@
+### Backwards-incompatible changes
+
+- Drop support for getting user metadata from OpenID Connect token claims. LDAP, for both user metadata and group membership, is now required when using an OpenID Connect authentication, including CILogon.
+- Drop support for LDAP groups without GIDs. Either Firestore GID assignment must be enabled or LDAP must contain a GID for each group. Groups without GIDs in LDAP will be ignored if Firestore is not enabled.
+- Retrieval of the UID and primary GID from LDAP is now enabled by default unless Firestore is enabled.
