+++ src/gafaelfawr/models/token.py
@@ -5,7 +5,6 @@ from __future__ import annotations
 from datetime import datetime
 from enum import Enum
 from typing import Any, Self
-from uuid import UUID
 
 from pydantic import (
     BaseModel,
@@ -18,20 +17,17 @@ from pydantic import (
 from safir.datetime import current_datetime
 from safir.pydantic import normalize_datetime
 
-from ..constants import GROUPNAME_REGEX, USERNAME_REGEX
+from ..constants import USERNAME_REGEX
 from ..exceptions import InvalidTokenError
 from ..util import normalize_scopes, random_128_bits
+from .userinfo import Group
 
 __all__ = [
     "AdminTokenRequest",
-    "CADCUserInfo",
     "NewToken",
-    "NotebookQuota",
-    "Quota",
     "Token",
     "TokenBase",
     "TokenData",
-    "TokenGroup",
     "TokenInfo",
     "TokenType",
     "TokenUserInfo",
@@ -149,65 +145,6 @@ class TokenType(Enum):
     """Access token for an OpenID Connect client."""
 
 
-class TokenGroup(BaseModel):
-    """Information about a single group."""
-
-    name: str = Field(
-        ...,
-        title="Name of the group",
-        examples=["g_special_users"],
-        min_length=1,
-        pattern=GROUPNAME_REGEX,
-    )
-
-    id: int | None = Field(
-        None,
-        title="Numeric GID of the group",
-        examples=[123181],
-        description=(
-            "Numeric GID may be unset, in which case the group still"
-            " contributes to determining scopes, but may be ignored by"
-            " services that require a GID. If Firestore is configured, a"
-            " numeric GID will be allocated by Firestore if left unset"
-            " when creating a token."
-        ),
-    )
-
-
-class NotebookQuota(BaseModel):
-    """Notebook Aspect quota information for a user."""
-
-    cpu: float = Field(..., title="CPU equivalents", examples=[4.0])
-
-    memory: float = Field(
-        ..., title="Maximum memory use (GiB)", examples=[16.0]
-    )
-
-
-class Quota(BaseModel):
-    """Quota information for a user."""
-
-    api: dict[str, int] = Field(
-        {},
-        title="API quotas",
-        description=(
-            "Mapping of service names to allowed requests per 15 minutes."
-        ),
-        examples=[
-            {
-                "datalinker": 500,
-                "hips": 2000,
-                "tap": 500,
-                "vo-cutouts": 100,
-            }
-        ],
-    )
-
-    notebook: NotebookQuota | None = Field(
-        None, title="Notebook Aspect quotas"
-    )
-
-
 class TokenBase(BaseModel):
     """Base information about a token common to several representations.
 
@@ -251,16 +188,20 @@ class TokenBase(BaseModel):
         examples=[1616986130],
     )
 
+    @field_serializer("created", "expires")
+    def _serialize_datetime(self, time: datetime | None) -> int | None:
+        return int(time.timestamp()) if time is not None else None
+
     _normalize_scopes = field_validator("scopes", mode="before")(
         normalize_scopes
     )
 
 
 class TokenInfo(TokenBase):
-    """Information about a token returned by the token-info endpoint.
+    """Information about a token.
 
-    This is all the information about the token that's stored in the
-    underlying database.  It includes some fields not present in Redis.
+    All information about the token that's stored in the underlying
+    database. It includes some fields not present in Redis.
     """
 
     token: str = Field(
@@ -318,11 +259,13 @@ class TokenInfo(TokenBase):
 
 
 class TokenUserInfo(BaseModel):
-    """The information about a user stored with their token.
+    """User metadata stored with the token.
 
-    If information is stored with the token, it overrides information from
-    other sources such as LDAP.  Fields that should be dynamically retrieved
-    from LDAP should be omitted or set to `None`.
+    Holds the user metadata generated by the authentication provider or the
+    admin token request and stored with the token, without any supplemental
+    information from LDAP or Firestore. Fields that are set will override any
+    data from extenral sources. Fields set to `None` will be looked up in LDAP
+    or Firestore.
     """
 
     username: str = Field(
@@ -361,23 +304,21 @@ class TokenUserInfo(BaseModel):
         ge=1,
     )
 
-    groups: list[TokenGroup] | None = Field(
+    groups: list[Group] | None = Field(
         None,
         title="Groups",
         description="Groups of which the user is a member",
     )
-
-    quota: Quota | None = Field(None, title="Quota")
 
     def to_userinfo_dict(self) -> dict[str, Any]:
         """Convert to a dictionary for logging purposes.
 
         This method converts only the `TokenUserInfo` portion of a token to a
         dictionary for logging purposes, excluding the ``username`` field
-        (which is logged separately).  It's used when logging creation of
-        new tokens to make a record of the user identity information included
-        in the token (as opposed to retrieved dynamically from other sources
-        such as LDAP or Firestore).
+        (which is logged separately). It's used when logging creation of new
+        tokens to make a record of the user identity information included in
+        the token (as opposed to retrieved dynamically from other sources such
+        as LDAP or Firestore).
 
         Returns
         -------
@@ -411,41 +352,6 @@ class TokenData(TokenBase, TokenUserInfo):
     """
 
     token: Token = Field(..., title="Associated token")
-
-    # These fields should not have to be reiterated since they're inherited
-    # from TokenUserInfo, but the Pydantic mypy plugin has a bug that loses
-    # inherited | None type information. See:
-    #
-    # https://github.com/pydantic/pydantic/issues/7398
-    #
-    # These duplicate definitions should be removed when the Pydantic mypy
-    # plugin is fixed.
-    name: str | None = Field(
-        None,
-        title="Preferred full name",
-        examples=["Alice Example"],
-        min_length=1,
-    )
-    uid: int | None = Field(None, title="UID number", examples=[4123], ge=1)
-    gid: int | None = Field(
-        None,
-        title="Primary GID",
-        description=(
-            "GID of primary group. If set, this will also be the GID of one of"
-            " the groups of which the user is a member."
-        ),
-        examples=[4123],
-        ge=1,
-    )
-    groups: list[TokenGroup] | None = Field(
-        None,
-        title="Groups",
-        description="Groups of which the user is a member",
-    )
-
-    @field_serializer("created", "expires")
-    def _serialize_datetime(self, time: datetime | None) -> int | None:
-        return int(time.timestamp()) if time is not None else None
 
     @classmethod
     def bootstrap_token(cls) -> Self:
@@ -594,7 +500,7 @@ class AdminTokenRequest(BaseModel):
         ge=1,
     )
 
-    groups: list[TokenGroup] | None = Field(
+    groups: list[Group] | None = Field(
         None,
         title="Groups",
         description=(
@@ -679,45 +585,3 @@ class UserTokenModifyRequest(BaseModel):
         ),
         examples=[1625986130],
     )
-
-
-class CADCUserInfo(BaseModel):
-    """User metadata required by the CADC authentication code.
-
-    This model is hopefully temporary and will be retired by merging the CADC
-    support with the OpenID Connect support.
-    """
-
-    exp: datetime | None = Field(
-        None,
-        title="Expiration time",
-        description=(
-            "Expiration timestamp of the token in seconds since epoch. If"
-            " not present, the token never expires."
-        ),
-        examples=[1625986130],
-    )
-
-    preferred_username: str = Field(
-        ...,
-        title="Username",
-        description="Username of user",
-        examples=["someuser"],
-    )
-
-    sub: UUID = Field(
-        ...,
-        title="Unique identifier",
-        description=(
-            "CADC code currently requires a UUID in this field. In practice,"
-            " this is generated as a version 5 UUID based on a configured"
-            " UUID namespace and the string representation of the user's"
-            " numeric UID (thus allowing username changes if the UID is"
-            " preserved."
-        ),
-        examples=["78410c93-841d-53b1-a353-6411524d149e"],
-    )
-
-    @field_serializer("exp")
-    def _serialize_datetime(self, time: datetime | None) -> int | None:
-        return int(time.timestamp()) if time is not None else None
