+++ tests/data/config/oidc-no-kids.yaml.in
@@ -27,3 +27,7 @@ oidc:
     - "voPerson"
   issuer: "https://upstream.example.com/"
   audience: "https://test.example.com/"
+ldap:
+  url: "ldaps://ldap.example.com/"
+  groupBaseDn: "dc=example,dc=com"
+  userBaseDn: "ou=people,dc=example,dc=com"
